% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_eRNA.R
\name{find_eRNA}
\alias{find_eRNA}
\title{Find eRNA (Enhancer RNA) Transcripts}
\usage{
find_eRNA(stg_path, ignore.strand = FALSE)
}
\arguments{
\item{stg_path}{Character. Path to StringTie GTF file}

\item{ignore.strand}{Logical. If FALSE (default), strand-specific matching}
}
\value{
GRanges object of eRNA transcripts
}
\description{
Find eRNA (Enhancer RNA) Transcripts
}
\examples{
\dontrun{
  eRNA <- find_eRNA("stringtie.gtf")
}

}
